TDD With MVC 5 and Entity Framework and Repository pattern/[C#]-TDD With MVC 5 and Entity Framework and Repository pattern/C#/TDD.ServicesLayer/Repository/StudentRepository.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Data;
using TDD.Entities.Models;
using TDD.Entities.DAL;
using System.Data.Entity;

namespace TDD.ServicesLayer
{
    public class StudentRepository : IStudentRepository, IDisposable
    {
        private SchoolContext context;

        public StudentRepository(SchoolContext context)
        {
            this.context = context;
        }

        public IEnumerable<Student> GetStudents()
        {
            return context.Students.ToList();
        }

        public Student GetStudentByID(int id)
        {
            return context.Students.Find(id);
        }

        public void InsertStudent(Student student)
        {
            context.Students.Add(student);
        }
         public int AddStudent(Student student)
        {
            context.Students.Add(student);
            context.SaveChanges();
            return student.ID;
        }

         
        public void DeleteStudent(int studentID)
        {
            Student student = context.Students.Find(studentID);
            context.Students.Remove(student);
        }

        public void UpdateStudent(Student student)
        {
            context.Entry(student).State = EntityState.Modified;
        }

        public void Save()
        {
            context.SaveChanges();
        }

        private bool disposed = false;

        protected virtual void Dispose(bool disposing)
        {
            if (!this.disposed)
            {
                if (disposing)
                {
                    context.Dispose();
                }
            }
            this.disposed = true;
        }

        public void Dispose()
        {
            Dispose(true);
            GC.SuppressFinalize(this);
        }
    }
}
